Move metrics db functions to dedicated store
